+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/ColumnSummaryPanel.java
@@ -18,7 +18,7 @@
  */
 package is.codion.swing.common.ui.component.table;
 
-import is.codion.common.model.table.ColumnSummaryModel;
+import is.codion.common.model.summary.ColumnSummaryModel;
 import is.codion.swing.common.ui.component.Components;
 import is.codion.swing.common.ui.control.Control;