+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityTablePanel.java
@@ -24,7 +24,7 @@ import is.codion.common.i18n.Messages;
 import is.codion.common.model.UserPreferences;
 import is.codion.common.model.condition.ColumnConditionModel;
 import is.codion.common.model.condition.TableConditionModel;
-import is.codion.common.model.table.ColumnSummaryModel;
+import is.codion.common.model.summary.ColumnSummaryModel;
 import is.codion.common.property.PropertyValue;
 import is.codion.common.resource.MessageBundle;
 import is.codion.common.state.State;